move player representation code
- misc changes so it uses the character type enum define instead of a typemap
